Update release to jammy stemcell

+++ acceptance-tests/run-local.sh
@@ -13,6 +13,8 @@ if [ "$(git status -s | wc -l)" -gt 0 ]; then
     exit 1
 fi
 
+FOCUS="$1"
+
 docker_mac_check_cgroupsv1() {
     # Force cgroups v1 on Docker for Mac
     # inspired by https://github.com/docker/for-mac/issues/6073#issuecomment-1018793677
@@ -37,4 +39,8 @@ pushd "$SCRIPT_DIR/../ci" || exit 1
 popd || exit 1
 
 # Run acceptance tests
-docker run --rm --privileged -v "$REPO_DIR":/repo -e REPO_ROOT=/repo haproxy-boshrelease-testflight bash -c "cd /repo/ci/scripts && ./acceptance-tests"
+if [ -n "$FOCUS" ]; then
+  docker run --privileged -v "$REPO_DIR":/repo -e REPO_ROOT=/repo -e FOCUS="$FOCUS" haproxy-boshrelease-testflight bash -c "cd /repo/ci/scripts && ./acceptance-tests ; sleep infinity"
+else
+  docker run --rm --privileged -v "$REPO_DIR":/repo -e REPO_ROOT=/repo haproxy-boshrelease-testflight bash -c "cd /repo/ci/scripts && ./acceptance-tests"
+fi
